+++ cards/moengage_cards_ios/CHANGELOG.md
@@ -1,5 +1,10 @@
 # MoEngage Cards iOS Plugin
 
+# Next Release date
+
+## Next version
+- Refresh cards on uniqueID set
+
 # 07-08-2024
 
 ## 3.0.1